admin

+++ user.php
@@ -1,7 +1,6 @@
 <?php
-if(!isset($_SESSION)){
     session_start();
-}    
+   
     include 'config/config.php';
 
     try {
@@ -9,11 +8,13 @@ if(!isset($_SESSION)){
 	} catch (Exception $e) {
 		die('Erreur : ' . $e->getMessage());
     }
+    var_dump($_SESSION);
     if(isset($_GET['id']) AND $_GET['id'] > 0){
         $getid = intval($_GET['id']);
-        $requser = $bdd->prepare("SELECT * FROM utilisateur WHERE id = ?");
-        $requser->execute(array($getid));
-        $userinfo = $requser->fetch();
+        $req = $bdd->prepare("SELECT * FROM utilisateur WHERE id = ?");
+        $req->execute(array($getid));
+        $user = $req->fetch();
+        var_dump($user['admin']);
         if(isset($_GET['supprimerUser']) AND !empty($_GET['supprimerUser'])){
             $supprimerUser = (int) $_GET['supprimerUser'];
             $deleteUser = $bdd->prepare('DELETE FROM utilisateur WHERE id = ? LIMIT 1');
@@ -22,7 +23,7 @@ if(!isset($_SESSION)){
         }
         if(isset($_GET['supprimerEvent']) AND !empty($_GET['supprimerEvent'])){
             $supprimerEvent = (int) $_GET['supprimerEvent'];
-            $deleteEvent = $bdd->prepare('DELETE FROM evenement WHERE id = ? LIMIT 1');
+            $deleteEvent = $bdd->prepare('DELETE FROM evenement WHERE id = ? LIMIT 1');var_dump($user['admin']);
             $deleteEvent->execute(array($supprimerEvent));
             header('Location: user.php?id=' .$_SESSION['id']);
         }
@@ -69,11 +70,11 @@ if(!isset($_SESSION)){
                 <section class="article">
                     <div class="range1">
                         <?php
-                            if(isset($_SESSION['id']) AND $userinfo['id'] == $_SESSION['id'] AND $_SESSION['id'] == 1){
+                            if(isset($_SESSION['id']) AND $_SESSION['admin'] == 1){
                         ?>
                             <a href="edit_profile.php?id=<?php echo $_SESSION['id']; ?>"><i class="fas fa-pen buttonsection"></i></a>
                         <?php
-                            }elseif(isset($_SESSION['id']) AND $userinfo['id'] == $_SESSION['id']){
+                            }elseif(isset($_SESSION['id'])){
                         ?>
                             <a href="edit_profile.php?id=<?php echo $_SESSION['id']; ?>"><i class="fas fa-pen buttonsection"></i></a>
                             <button class='deletebutton' onclick="ConfirmDelete()"><i class="fas fa-trash buttonsection"></i></button>
@@ -81,10 +82,10 @@ if(!isset($_SESSION)){
                             }
                         ?> 
                     </div>
-                        <img src="user/avatar/<?php echo $userinfo['avatar']; ?>" alt="image user" class="imguser" width='150'>
-                        <h2 class="titre-h2"><?php echo $userinfo['pseudo']; ?></h2>
-                        <h5 class="email"><?php echo $userinfo['mail']; ?></h5>
-                        <?php if($_SESSION['id'] == $_GET['id'] AND $_SESSION['id'] == 1){?>
+                        <img src="user/avatar/<?php echo $user['avatar']; ?>" alt="image user" class="imguser" width='150'>
+                        <h2 class="titre-h2"><?php echo $user['pseudo']; ?></h2>
+                        <h5 class="email"><?php echo $user['mail']; ?></h5>
+                        <?php if($_SESSION['id'] == $_GET['id'] AND $_SESSION['admin'] == 1){?>
                             <p>Administrateur</p>
                         <?php }else{ ?>
                             <p>Utilisateur</p>
@@ -92,7 +93,7 @@ if(!isset($_SESSION)){
                 </section>
             </div>
             <?php
-                if(isset($_GET['id']) AND $_GET['id'] == $_SESSION['id'] AND $_SESSION['id'] == 1){
+                if(isset($_GET['id']) AND $_GET['id'] == $_SESSION['id'] AND $_SESSION['admin'] == 1){
                     $getidAdmin = intval($_GET['id']);
             ?>
             <div class='admin'>
